third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-center-vrl-vrl-expected.txt

This is a testharness.js-based test.
[FAIL] .target 4
  assert_equals: \n<div class="container">\n  <div class="anchor"></div>\n  <div class="target" style="bottom: 20px;" data-expected-height="80" data-offset-y="0"></div>\n</div>\noffsetTop expected 0 but got 20
[FAIL] .target 8
  assert_equals: \n<div class="container">\n  <div class="anchor"></div>\n  <div class="target" style="top: 10px; bottom: 20px;" data-expected-height="70" data-offset-y="10"></div>\n</div>\noffsetTop expected 10 but got 30
Harness: the test ran to completion.
